[#58] refactor: Spy 객체를 재사용하도록 파일 분리하고 재적용
