+++ iOS/SignupApp/SignupApp.xcodeproj/xcuserdata/kimdo.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -10,5 +10,18 @@
 			<integer>0</integer>
 		</dict>
 	</dict>
+	<key>SuppressBuildableAutocreation</key>
+	<dict>
+		<key>5A834507242A301B00F20FAA</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>5A83451D242A301E00F20FAA</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+	</dict>
 </dict>
 </plist>
